Stopped rounding the meters' decimal digit away at the telemetry publisher
Every caller already passed full-precision BigDecimal — flow_reading_table holds the readings as NUMERIC and the correction paths derive their volumes by subtracting two of them — so toInt() in TelemetryEventPublisher was the single lossy step between the source of truth and the warehouse. It rounded HALF_UP to whole cubic metres, discarding up to 0.5 m3 per reading, which is up to 1000 L on the daily delta analytics derives from two of them.

MeterReadingEvent's readings and WaterQuantityEvent.waterQuantity are now BigDecimal and both publishers pass their arguments through untouched. The reason-only publishers send BigDecimal.ZERO. toInt(BigDecimal) has no callers left and is gone; toInt(Long) stays for the id narrowing, and toConfidenceInt keeps its own rounding, which is a genuine normalisation to a whole percentage.

+++ backend/telemetry-service/src/main/java/org/arghyam/jalsoochak/telemetry/dto/event/MeterReadingEvent.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -15,8 +17,13 @@ public class MeterReadingEvent {
     private Integer tenantId;
     private Integer schemeId;
     private Integer userId;
-    private Integer extractedReading;
-    private Integer confirmedReading;
+    /**
+     * The readings exactly as {@code flow_reading_table} holds them — {@code NUMERIC}, decimal digit
+     * included. They used to be rounded to whole cubic metres on the way out, which cost the warehouse
+     * up to 1000 L on every daily volume it derived from them.
+     */
+    private BigDecimal extractedReading;
+    private BigDecimal confirmedReading;
     private Integer confidence;
     private String imageUrl;
     private String readingAt;
